Final. - Updated users, regions, and institutions migration tables.

+++ resources/views/sysadmin/institutions/index.blade.php
@@ -33,13 +33,19 @@
                     <td>{{ $institution->dean_phone }}</td>
                     <td>{{ $institution->region->name }}</td>
                     <td>
-                        <a href="institutions/{{ $institution->id }}/edit" title="Edit"><span class="glyphicon glyphicon-edit" aria-hidden="true"></span></a>
+                        @if($institution->name != 'None')
+                            <a href="institutions/{{ $institution->id }}/edit" title="Edit"><span class="glyphicon glyphicon-edit" aria-hidden="true"></span></a>
+                        @endif
                     </td>
                     <td>
-                        <a href="institutions/{{ $institution->id }}/delete" title="Delete"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a>
+                        @if($institution->name != 'None')
+                            <a href="institutions/{{ $institution->id }}/delete" title="Delete"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a>
+                        @endif
                     </td>
                     <td>
-                        <a href="institutions/{{ $institution->id }}" title="View"><span class="glyphicon glyphicon-eye-open" aria-hidden="true"></span></a>
+                        @if($institution->name != 'None')
+                            <a href="institutions/{{ $institution->id }}" title="View"><span class="glyphicon glyphicon-eye-open" aria-hidden="true"></span></a>
+                        @endif
                     </td>
                 </tr>
             @empty
